FEM - Rework from week 6
Used position fixed for the pop up on larger screens. However, I then had to use JavaScript to get the position of the share button to set the position of the pop up.
I also removed some excessive divs on the avatar section.

One issue with Firefox is the pseud elements on the pop up have a white line between them - Not an issue on Chrome or Opera.